Added question create functionality

+++ app/src/main/java/com/example/educationalchallenge/dto/ProfileResponse.java
@@ -1,11 +1,11 @@
 package com.example.educationalchallenge.dto;
 
 public class ProfileResponse {
-    public int id;
+    public Long id;
     public String username;
     public String createdDate;
-    public int level;
-    public int exp;
+    public Integer level;
+    public Integer exp;
     public String role;
-    public int quizzesSolved;
+    public Integer quizzesSolved;
 }
